py-pyobjc: update to 8.4.1
I elected to drop most of the patching of the build infrastructure,
and just use PEP-517 on each directory instead. Since the tools don't
actually support multiple directories or wheels, I added two small
scripts to do that. Hopefully this should prove more reliable and
maintainable in the long run.

Closes: https://trac.macports.org/ticket/64766