gcc-devel: improve header search locations
Setting the local prefix to the MacPorts `${prefix}` does little good
since the sysroot is prepended.
The same reasoning applies to the default value of `/usr/local`.
Therefore, do not set the local prefix.

Since the `-stdlib=` uses the system libc++, it should also use the
corresponding system headers.
When the sysroot is included in the path, the resulting compiler
respects the environment variable `SDKROOT`.