slum: Switch to fork, update, fix arm64 build
Upstream for slurm is dead. Switch to a fork that seems to be maintained
by some debian people and update to the current state (note that the
published releases are very old).

This fixes the build on arm64, which would otherwise require
a autoreconf because config.sub does not understand the arm64 machine.