openssl3: Drop patches that may no longer be needed
The build fails when compiling for i386 because the patches don't apply.
It seems https://github.com/openssl/openssl/pull/18056 with
-DBROKEN_CLANG_ATOMICS fixes this.

There is a patch upstream at
https://github.com/openssl/openssl/pull/16584, but carrying that
downstream would mean continued downstream maintenance, and since this
only affects a platform I cannot test, it is likely to break again with
the next update.

Closes: https://trac.macports.org/ticket/65111