nginx: Make variants work again
The fix for https://trac.macports.org/ticket/65150 had the unintended
consequence that the configure arguments from all variants were ignored.

In Tcl, "#" only starts a comment at the beginning of a line, so the "#
pcre2 breaks the lua module" comment was added to the configure args
literally. Configure args for the variants were added after that, but
because the shell recognizes "#" as a comment character anywhere in the
line, all those added configure args were ignored.