libcxx: adjust emulated_tls logic
Don't enforce compiler choice for emulated_tls until the variant is
confirmed to be set (it can be unset by the user even though it is a
default variant).

Default to +emulated_tls even with no suitable compiler active if the
current version of libcxx is already installed with +emulated_tls. It
doesn't need to be built in that case, but it may need to be activated,
and this ensures that it is activated with this variant.