saml2aws: update to 2.35.0
Switch from go.vendors to allowing Go to download source.

This port has several dependencies which don't play well with the
golang portgroup and requires lots of manual massaging on every upgrade.
One dependency has now changed GitHub organisation, but go.sum doesn't
reflect that, causing an incompatibility in the portgroup: if you fetch
from the new location the port fails to build because it expects the old
path; however the portgroup fails to download when using the old
location.