ffmpeg: prefer libsmbclient from samba4 instead of samba3
ffmpeg+gpl3’s libsmbclient dependency may be satisfied by either samba3
or samba4. This updates its preference to samba4.

The port’s revision is unchanged because no port rebuild is required. If
samba3 is installed, it will continue to satisfy this dependency. As
samba3 and samba4 conflict, there is no ambiguity within a MacPorts
installation, and ffmpeg will continue to depend on whichever samba
version is installed absent user intervention.

This also updates notes to refer to libsmbclient by the correct name.