getopt: replace with util-linux
The util-linux version of getopt was synced with the last-released
version of the standalone getopt (1.1.6, 2014-11-24) at b5f3078e1474
(2014-12-08). Since then, there have been no changes to the standalone
getopt, thus it appears unmaintained. There have been changes to the
copy in util-linux, which is still actively maintained.

This makes the util-linux port build and install getopt, updates
dependents of the existing getopt port to instead depend on util-linux,
and marks the existing getopt port as replaced by util-linux.