rawtherapee: fix broken port due to source code unable to find GtkMacIntegration lib files
This commit fixes the rawtherapee port, which became broken due to recent changes in MacPorts that were needed to segregate the various subports of gtk-osx-application-xx.

The fix is implemented as a CMake find module, which will, in turn, utilize PkgConfig to locate the GtkMacIntegration library. Since this was basically a complete rewrite of the ${worksrcpath}/cmake/modules/FindMacIntegration.cmake file, I instead created a new file with the name 'FindGtkMacIntegration.cmake' (which I believe to be the more proper name for the file anyway). At some point, this file should be submitted to upstream so that it gets incorporated into the RawTherapee code base.

Closes: https://trac.macports.org/ticket/65283