qt*-qtbase: improve SDK detection
The qt*-qtbase ports attempted to build a valid SDK name as
“macosx${configure.sdk_version}”. This candidate SDK was validated by
checking whether “MacOSX${configure.sdk_version}.sdk” exists in the SDKs
directory within the developer directory.

In 9887e90d69f4, identical logic in the qmake5 port group was revised to
instead check for a valid SDK based on ${configure.sdk_version} by
running “xcrun --sdk macosx${configure.sdk_version} --show-sdk-path”.
This was later revised by be588d4c1b8b to use “--find ld” instead of
“--show-sdk-path”, which is a newer option not honored by all versions
of xcrun.

This change brings the improved logic from the qmake5 port group (and
later updates) to the various qt*-qtbase subports, which build via the
Qt configure script and not qmake, and do not otherwise benefit from the
improvements already made to the qmake5 port group.

This is necessary because unlike other recent versions of Xcode, Xcode
14.0b1 14A5228q contains a MacOSX13.sdk symbolic link in addition to the
expected MacOSX13.0.sdk symbolic link. This means that the path
existence check for MacOSX${configure.sdk_version}.sdk succeeded in this
configuration when configure.sdk_version was 13, as it will be by
default on macOS 13. This resulted in Qt being configured with an SDK
identifier of “macosx13”. Although the MacOSX13.sdk symbolic link was
present, xcrun and other Xcode tools do not recognize “macosx13” as a
valid SDK identifier, causing a build configured in this way to fail.

The expectation based on other recent versions of Xcode (since Xcode 12
and the macOS 11 SDK) is for the existence check to fail, causing the
build to fall back to the generic “macosx” SDK. By instead using xcrun
with a candidate SDK name (“xcrun --sdk macosx13”), these ports can now
check to see whether the SDK identifier to be given to the Qt build is
considered valid by the Developer Tools installation. This is much more
accurate than the path existence check.

The qt*-qtbase build winds up encoding the SDK identifier into its
installed files for its own use during qmake-based builds as
QMAKE_MAC_SDK, so it’s most desirable for MacPorts’ purposes to track
the SDK major version when possible, but it’s absolutely undesirable to
track the SDK minor version, as it will cause qmake to fail if a
Developer Tools update changes the SDK’s minor version. These minor
version bumps have occurred in recent Developer Tools updates that
identify SDKs by both major and minor version numbers. Because
QMAKE_MAC_SDK may only be set to a short identifier (“macosx10.15” or
“macosx”) and not a pathname, there is no way to currently identify such
SDKs to the tools by just a major version number. The only remaining
option is to specify the SDK as the most generic “macosx”. However, if
the developer tools do eventually regain the capability to properly
specify an SDK by just major version (“macosx13”), the strategy
implemented here will allow these ports to take advantage of it.