64tass: Install docs and syntax
Add a post-destdir action to install the missing docs and syntax files.
The 64tass Makefile's install target does not install the NEWS file nor
the plain-text README. The Makefile also does not install the 64tass
syntax files. The Makefile does already install the LICENSE* files, so
we don't need to copy LICENSE* files using the Portfile.