gnupg2: Update to 2.2.36
See release announcement for 2.3.7, which explains the CVE:
https://lists.gnupg.org/pipermail/gnupg-announce/2022q3/000474.html.
There does not seem to be a release announcement for 2.2.36.

I could not find the SHA-1 hash on the website as documented in the
Portfile, but I did verify the signature. I kept the SHA-1 digest
anyway, since this is an openmaintainer security update and not the
place for cleanup.

CVE: CVE-2022-34903