fluidsynth: update to 2.2.8, revbump known dependents
This closes Trac ticket #63501. This update to fluidsynth introduces
API and ABI breakage, compilation of all ports except audacious, lmms,
and VLC2 succeeded in local building (with trace mode) using this
version of fluidsynth. Here is a summary of changes to dependents.

* audacious-plugins
  - revbump
  Note: audacious-core fails compilation due to missing the
  dependency gdk-x11-2.0. This does not appear available on MacPorts.
  Moreover, the port maintainer states in Trac ticket #63501 to not
  treat compilation issues "as a blocker" for updating fluidsynth.

* denemo
  - revbump
  Note: this port successfully compiles when building locally, refer
  to GitHub ticket #14510

* gstreamer1-gst-plugins-bad
  - revbump

* libVLC2
  - see VLC2 below

* lmms
  - revbump
  Note: this port fails to compile due to encountering the issue
  described in Trac ticket #58009; this is NOT related to fluidsynth.

* scummvm
  - revbump

* scummvm-devel
  - revbump

* qsynth
  - revbump

* qtads
  - revbump

* VLC2
  - fix modeline for Emacs
  - revbump
  - remove fluidsynth dependency
  - disable fluidsynth in configuration
  Note: this port fails to compile locally, but the issue is NOT
  related to fluidsynth.

As a final note, apply this commit AFTER the pull request in GitHub
ticket #15565 is accepted. I decided to not revbump dosbox-x because
local testing suggests compatibility with fluidsyndth 2.2.8, and I do
not want to introduce a merge conflict when dosbox-x is updated.