clisp: update to 2.50.0
This updates the Portfile to pull in from the GNU CLISP GitLab
repository, where current development of GNU CLISP continues. While a
formal release has not been made, Bruno Haible (the maintainer of the
software) has cut out a "clisp-2.50.0" branch for those wanting a
release of GNU CLISP. The commit we use is the HEAD of this branch.