watchman: update to 2022.08.08.00
The recent update to folly broke building watchman. However, a simple
upgrade doesn't work, as they added yet another dependency to
something; their Eden SCM. Eden isn't in MacPorts, and I _really_
don't care about that one, so I disabled it -- which allowed pruning
the list of dependencies a bit. But they moved some code to rely on
it anyway, so I reverted those commits.

Since the update to Folly broke this, I added a comment to the port.