java PG: fix detection of arm
`${configure.build_arch}` and `[option configure.build_arch]` both
return an empty string here.

This lead to using default fallback on clean system which is `openjdkXY`
which probably won't be build on arm64.

To fix it I've migrated to `${os.arch}` which is available and which has
value `arm` instead of `arm64`.