p5-graphviz: update to version 2.25 (again)
A second update to the same version but with different content by
two different authors. Luckily the second update uses a different
extract.suffix. By fetching by author and specifying the corresponding
extract.suffix, the usual stealth update situation is avoided.

See the new pod for a discussion of how this came about.