zsh: Don't try to preprocess zshall.1 on Ventura
Preprocessing the `zshall.1` man page fails on macOS 13 Ventura because
Apple no longer ships a utility named `soelim`.  As I understand it,
this is because Ventura ships `mandoc` instead of `groff`.  Fortunately
for us, `mandoc` does not seem to require our preprocessing in the first
place, so there's no reason to attempt it.

Revbump because users with the `groff` port installed were able to build
this port successfully, but their `zshall.1` is different from the one
this revision produces.

Closes: https://trac.macports.org/ticket/65718