py-notify-python: fix missing python 2.7 symbols
pkg-config is used to create all the linker flags but misses
'-lpython2.7' which is required to fix the error in the ticket.
It's not possible to change pkg-config output from this port's
context (and dangerous).

It's not possible to patch the generated src/Makefile because
in it are paths relative to a specific MacPorts install.

Appending to the variable in configure is therefore the simplest
solution.

Closes: https://trac.macports.org/ticket/65092