gawk: reenable arm64 build
Upstream gawk has added a “PMA” (persistent memory allocator) feature
that unfortunately requires gawk to be built as a non-PIE
(position-independent executable). Non-PIE is not supported at all on
mac-arm64, so the upstream default as of
https://git.savannah.gnu.org/cgit/gawk.git/commit?id=0d9727bcd312da8555ca59942790adfee5fcafea
is to forcibly build gawk as mac-x86_64, requiring it to run under
Rosetta binary translation on mac-arm64. Informed opinion: the
upstream’s approach is misguided for multiple reasons.

The recent update of the gawk port to 5.2.1 at 42141802c346 honored this
upstream default, causing problems for mac-arm64 systems without Rosetta
installed, and problems upgrading the gawk port to this new version in
the likely case that the existing gettext-runtime dependency was not
installed as its universal variant.

To reenable the arm64 build of gawk, configure it to disable its PMA
feature, allowing it to be built as PIE and thus as a proper arm64
executable.

As non-PIE is deprecated when targeting macOS ≥ 12 on any architecture,
disable gawk’s PMA feature in that case as well.

Closes: https://trac.macports.org/ticket/66327