ci: Fix errors during step summary creation
It seems the linebreaks inside of double quotes are not correctly
rendered, but since '\n' is successfully used further down with tr -d,
that should work instead.

Remove the tail of the dependency progress log, since it seems to no
longer be available at the point where we are trying to print it.