ntpsec: Fix build failure with clang 14.
This was actually just a test failure in a corner case, but since the
tests are run as part of the build, the build was failing.  The build
originally worked on 12.x Monterey, but failed after upgrading to
Xcode 14, and building on 13.x Ventura never worked.

There was already a one-line upstream fix for this, which is
backported here.

Also, the Python programs lacked a versioned shebang line, causing
them to fail whenever the default Python wasn't python27.  This also
now fixed.

Since installed content is changed, this includes a revbump.

This doesn't bother to fix the minor lint warning about a duplicate
pkgconfig dependency.

As usual, the patchfile is derived from a GitLab fork, and the overall
diffs can be seen at:

    https://gitlab.com/fhgwright/ntpsec/tree/macports_1_2_1r4

TESTED:
Tested (including building the usual set of variant combinations) on
10.4-10.5 ppc, 10.4-10.6 i386, 10.6-10.15 x86_64, and 11.x-13.x arm64.