py-build: update to 0.9.0
The upgrade addresses the issues with Colorama & trace mode. The
issues with `importlib_metadata` seem to persist, but only for certain
older versions of Python as far as I can tell. To work around this, we
simply add it as (sort of) redundant dependency.

See https://trac.macports.org/ticket/65109