qt6: bump qtbase revision
The PR #16831 bumped qtbase revision from 0 to 1
but because of 41000ab0296c57c01f3aba71bcdf5aa52b68ceb3 which was merged before,
the revision was not bumped and hence, qtbase was not rebuilt with the changes
of #16831.

So bump it to revision 2 (and dependants also). This should fix build failure
https://build.macports.org/builders/ports-13_x86_64-watcher/builds/1010