qt6: remove useless & confusing cmake symlinks
These symlinks are not useful IMO.

- qmake shall find the qt6 installation it belongs to correctly
- when building a Qt application with cmake, one should actually use
  qt-cmake which will also find the qt6 installation it belongs to.

The important thing is to call the qmake & qt-cmake executable from their
installation location. Eg /opt/local/libexec/qt6/bin

especially, qt-cmake shouldn't be called through a symlink

If these symlinks are present and configuration is made with cmake (and not
qt-cmake) one may have errors reported at configure time saying that some Qt
components/tools are not installed. Actually, this is due to heavy use by Qt
of CMAKE_CURRENT_LIST_DIR in the cmake files. Changing this will mess the Qt6
installation. It doesn't resolve to the realpath but to the path to the
symlink.

Configuration should be made with qt-cmake (or if one wants to use cmake, one
has to add additional parameters to the cmake call) or qmake.