py-serial: Add py311 subport, + additional fixes.
Also corrects the py-setuptools dependency, which is needed at
runtime for the included programs.  Because this affects the
runtime dependency, it includes a revbump.

Also restores the py34-py36 subports, which were removed without
maintainer approval.  The upstream code supports 2.7 and 3.4+.  These
subports are a known_fail on arm (handled externally), but OK where
python34-36 exists.

Also cleans up long Portfile lines in the long description.

NOTE:
Testing turned up bugs in python36 on 10.4 ppc, and python37 on
10.4-10.5 ppc, 10.5 i386, and 10.5 x86_64.  There also seem to be bugs
in setuptools in certain cases which cause trace-mode builds to fail.
None of these issues is believed to be a problem in this port itself.

TESTED:
Tested the included programs for all available Python versions on
10.4-10.5 ppc, 10.4-10.6 i386, 10.5-10.15 x86_64, and 11.x-13.x arm64.
Tested actual serial communication in a subset of cases.