various: set new extract.rename to yes if it exists
This feature is in current master and will be part of the next MacPorts release. When set to yes, it tries to rename the extracted tarball top-level directory to what the Portfile specifies, if different. Some ports require this while others do not. Here are 4 that I've come across in my recent work; I'm sure there are more.