ghc: Clean up end of checksums statement
This change is likely a no-op, other than fixing syntax highlighting in
the editor. I'm making this change because the last build of GHC on
x86_64 Monterey failed because it ran out of disk space, and our
buildbot will not attempt builds of dependencies of GHC (e.g., pandoc)
on this platform until the checksum of the Portfile changes or somebody
manually deletes the cache entry that marks this version of the port as
failed.