asciidoctor: Revert to last fully working version of Ruby (3.0)
Ruby 3.1 doesn't currently build for 10.4-10.5 x86, or 10.8-10.9, and
fixing that doesn't appear to be a simple matter of applying the same
fixes as for 2.7 and 3.0.  Ruby 3.0 now builds for all OS versions and
CPU architectures, making it a more compatible choice, particularly
given that asciidoctor isn't fussy about the Ruby version.

Also adds myself as a maintainer, mainly to keep dependency changes
from slipping through in the future.

Closes: https://trac.macports.org/ticket/66409

TESTED:
Built gpsd and ntpsec docs on OSX 10.4-10.5 ppc, 10.4-10.6 i386,
10.5-10.15 x86_64, and 11.x-13.x arm64.