gpsd: update to 3.25
TESTED:
Tested (including building the usual set of variant combinations and
running the tests) on 10.5 ppc, 10.5-10.6 i386, 10.5-10.15 x86_64, and
11.x-13.x arm64.
Some variant combinations were excluded in some cases due to issues
with the dependencies, though it's expected that they'd work if the
broken dependencies were fixed.
On 10.5, the python37 variant was excluded since python37 itself is
broken (segfault from ctypes).
Known not to build on 10.4.