lua: move complexity from patch files to Portfile
By minimizing patch files and changing the behavior of the various
subports in the Portfile, it will hopefully be easier not only to see
what is happening but also to change it if necessary.

Revbump is needed since full use of MacPorts flags via makefile PG
can change the installed files.