openssl*: Fix for 3+ universal archs
The merger_dont_diff setup for the >2 archs case was never updated to
reflect the new prefix with the move to libexec.  It was also never
updated for openssl3 to reflect the upstream naming change.  That
caused a broken opensslconf.h (configuration.h for openssl3) to be
created in the >2 archs case, with no reported error.

This in turn caused Python +universal to build "successfully" with the
hashlib and ssl extensions missing.

Since this case is fairly rare (and not even allowed in 10.6+), it
doesn't justify a revbump, even though it should technically get one.

For similar reasons, revbumping dependents isn't justifiable, though
a note is provided suggesting rebuilding them (only in the relevant
case).

TESTED:
Built +universal on 10.5 x86_64 with x86_64, i386, and ppc as
universal_archs.  Resulting opensslconf.h (or configuration.h) setup
is now reasonable.  With the openssl* updates, built python26-27
+universal and python32-35 +universal, and verified that the hashlib
and ssl imports work in all three architectures.  Currently,
python36-311 don't build +universal in this configuration, for
unrelated reasons.
The normal builds of openssl11 and openssl3 were also tested on the
full range of systems, due to the recent updates.