soprano: disable java backend
the JavaVM requirement is a problem, as there is no current universal JavaVM for arm64/x86_64
and the kde ports can't build as arm64. So they would need an x86_64 JavaVM installed if there
is one. Any JavaVM the user has installed on an arm Mac is unlikely to be Intel arch.
just disable the java backend to allow the kde4 ports to build.

closes: https://trac.macports.org/ticket/66783
closes: https://trac.macports.org/ticket/45745