ci: Make portlist a string not an array
There does not seem to be any need to treat portlist as an array in this
step since we only ever get the entire array here, and in subsequent
steps we already treat it as string. Getting the entire array fails if
there are no elements (i.e. if the commit only modified files that are
not port files).

See: https://trac.macports.org/ticket/62071