audacity: overhaul and upgrade
This upgrade and overhaul of the Audacity Portfile hopefully adresses the
building issues currently affecting it.
- port:audacity-legacy (v2.4.2) is renamed to audacity2-legacy and is the last
version to provide the old project format and (presumably) compatability with
older OS versions.
- the current port:audacity (3.0.2.129) is renamed to audacity3-legacy; it
(presumably) provides compatibility with OS X 10.7 and 10.8 .
- port audacity is upgraded to git/HEAD (3.2.5.x; development goes quickly ATM).
It provides a port:audacity-wxWidgets subport that is compulsory on 10.9 and
10.10; on newer OS versions it will also build against port:wxWidgets-3.2 .
- port:audacity-wxWidgets is a pure build dependency

Also relinquishing my maintainership.

Fixes: #17682
Closes: https://trac.macports.org/ticket/66801