gmtl: Fix build issues
Add patches to fix build failures when using clang, when using the
latest version of scons, when scons is using python 3, and to prevent
the creation of the unwanted file gmtl-0.6.1.tar.gz and to install the
file gmtl.pc.

Add runtime dependencies on flagpoll and python which are used by
gmtl-config. flagpoll is currently incompatible with python 3 and uses
python27 so use the same python here to reduce the number of pythons
a user needs to install.

Specify absolute path to scons.

Specify prefix at build time so that build output does not give the
erroneous impression that files will be installed in /usr/local.

Specify empty directories for Boost and CppUnit at build time so that
any Boost or CppUnit that might be installed in /usr/local is not
inadvertently used.

No need to specify destroot.cmd or destroot.target; the defaults work.

Add modeline.

Modernize checksums.

Set platforms to any and set supported_archs to noarch because this port
doesn't install anything architecture- or OS-specific.

Add GitHub handle to maintainers line.

Install documentation files.

Indicate license. The software claims to be licensed LGPL-2.1+ but there
is also a license addendum that imposes additional restrictions and the
LGPL does not permit the imposition of additional restrictions.

Rewrite master_sites to avoid redirects.

Closes: https://trac.macports.org/ticket/57421