asciidoc: Fix documentation install directory
The default for docdir is ${datarootdir}/doc/${PACKAGE_TARNAME}, and
while these have the values we expect them to have in the configure
script, this variable is left unexpanded and ${PACKAGE_TARNAME} is not
defined in the Makefile after Makefile generation, ends up empty, and
the documentation is installed in ${datarootdir}/doc. This is clearly
a bug, but I'm not sure whether it is in autoconf or asciidoc.

Closes: https://trac.macports.org/ticket/67081