autoconf: remove gettext dependency
Autoconf itself doesn't need gettext, and the subset of ports that need
to run autopoint when regenerating their autotools build system should
list it in depends_build. This should usually be the case already,
since if the gettext programs are needed to generate the configure
script, they are also needed when running it.

Fixes: https://trac.macports.org/ticket/67085