Revert "p5-dbd-pg: add support for building with postgresql15"
This reverts commit e7bc73bf69e8c26061b5701819f226147d1d53df which
attempted to generate variants on the fly rather than just defining
them explicitly. The result caused a parsing error that interpreted
the generated variants conflicts clause as declaring conflicts with
other postgresql ports rather than this port's other postgresql variants.

In reality, the various postgresql ports do not conflict with each other
and can be installed in parallel as they are installed in separate
versioned directories. Only this port's variants conflict with each other
since the port can only be built with one version of postgresql at a time.

Going back to explicit variant definitions that are readable and work as desired.