clang-12: remove flang/mlir
on all the other clang builds, flang is a separate
port. on clang-12 we put flang in as part of the
clang build, but this is a problem for i386, as
flang does not build i386, at present at least.

remove the flang/mlir build from clang-12 to allow
clang-12 to build on 32bit SnowLeopard.

if there is interest, a flang-12/mlir-12 port could be
added to this Portfile. However, flang-12 is still not
very functional at this point in time and so there is
likely neither interest nor need for it.