clang_dependency PG: switch to clang-11-bootstrap
Instead of blacklisting all clangs I've enforced to use
clang-11-bootstrap on all systems which ships with something older which
can't be used by system to build port which is kept inside clang
bootstrap cycle.

Anyway, I've replaced compiler only if desire one isn't available and if
desired was MacPorts clang in accepted list of versions.

This dramatically decrease number of compilers which needs to be build
on old system.

Thus, clang-11-bootstrap contains all necessary patches to be able to
build LLVM/Clang up to 15 (tested), and probably 16 (untested).

[skip ci]