ntpsec: Add variants for Python versions.
This implements pythonXY variants to allow choosing the Python
version.  The default remains +python27 for now, as noted.

Having subports for each Python version would be better, but is more
work, so for now only one version at a time can be selected.

Closes: https://trac.macports.org/ticket/63955

Also removes the erroneously installed 'runtests' program.

Also removes the unnecessary dependency on py-test.

This gets a revbump due to the 'runtests' removal, as well as due to
the registry update for the new variants.

TESTED:
Built and ran tests with all variants on 10.4-10.5 ppc, 10.4-10.6
i386, 10.6-10.15 x86_64, and 11.x-13.x arm64.  Successful except where
pythonXY was unavailable or broken on the OS/CPU combination.