afscompress: Update to 1.7.3
Set supported OS versions in platforms variable rather than doing it
manually.

Remove "universal_variant yes" which is the default.

Remove modeline because the whitespace does not conform to it.

Use "github.tarball_from archive" so that the worksrcdir does not have
to be changed with each update as of MacPorts 2.8.1.

Closes: https://trac.macports.org/ticket/67650