Use PortGroups to avoid hardcoding values
This allows the use of any compiler that supports the C++20 standard.
Since MacPorts does not yet know what that standard is
(see https://trac.macports.org/ticket/62677), I am also blacklisting
versions of Clang and GCC that do not work to build btop.

Note that this currently fails with GCC on my machine, since it attempts
to use the system libc++, which does not support the features of C++20
that btop requires.