cosmoplayer: Fix destroot failure
They typo'd their own project name... in the worksrcdir... in 1999...

We had originally fixed this when the port was added in 38d64214, but
that conflicted with MacPorts 2.6.0's automatic worksrcdir symlinking so
the fix was removed in 8d5e8a87. It broke again when MacPorts 2.8.1
removed the automatic worksrcdir symlinking.