ntpsec: update to 1.2.2a, fixing crash bug
This release is a focused update to 1.2.2, addressing an issue related
to (but not quite the same as):

     https://gitlab.com/NTPsec/ntpsec/-/issues/794

Also fixes the livecheck for patch updates (like this one), by
allowing an optional single letter after the version number.

As usual, this includes the patches for compatibility with
macOS<10.13, which can also be seen (more readably) at:

    https://gitlab.com/fhgwright/ntpsec/tree/macports-1_2_2a

The patchfile is renormalized for the current mtimes, but not
substantively changed.

Note that 10.5 x86_64 remains broken and is disabled, as is the
untested 10.4 x86_64.

TESTED:
Built and ran tests with all variants on 10.4-10.5 ppc, 10.4-10.6
i386, 10.6-10.15 x86_64, and 11.x-13.x arm64.  Successful except where
pythonXY was unavailable or broken on the OS/CPU combination.