tigervnc: Make vncviewer open in low-DPI mode
TigerVNC Viewer is not compatible with high-DPI mode and its Info.plist
says so but if the application is launched via the vncviewer symlink
this bypasses the Info.plist and the app opens in high-DPI mode which
doesn't work right. Change the symlink to a wrapper script that runs the
real executable; this seems to allow the OS to use the Info.plist file
and open in low-DPI mode.

Closes: https://trac.macports.org/ticket/67939