zig: update to 0.11.0
See https://trac.macports.org/ticket/65333.

Note:

The use of "PortGroup cmake 1.1" would lead to a dependency on the
existence of `${prefix}/share/cmake/Modules`, but this directory may not
exist on some machines (such as the CI). We workaround this by
piggybacking on a directory that is known to exist due to the build
dependency on LLVM by this port.