muniversal: Remove quoted arch flags correctly
When removing architecture-specific flags, account for the fact that
sometimes (as in php-config) they are in quotation marks.

Also, do not collapse multiple spaces into one because that could result
in erroneous presentation of output (as in php-config) or an
IndentationError (in the case of python scripts).

Fixes and partially reverts c9a6e03fc195486bbc91148a44393ae9f333f109.

Closes: https://trac.macports.org/ticket/64758
Closes: https://trac.macports.org/ticket/67898