git-extras: prevent `git-extras update` error
#### Description

`git-extras` should not allow the `update` command whne managed by
MacPorts. There is a `brew-release.patch` file in the root of the
repository that is used when they prepare the Homebrew release.

I applied that file, changed the references from Homebrew to MacPorts,
and created a new patch file to add to the `devel/git-extras` port.

This prevents the following error:

```console
$ git-extras update
/dev/stdin: line 13: cd: $'/var/folders/c6/k_j3j3w11590l8x6t8rwxw200000gn/T/-d.CpB5Qci2\ngit-extras-install.9TwCl44Fql': No such file or directory
```

It prints instead

```console
$ git-extras update
This git-extras installation is managed by MacPorts.
If you'd like to update git-extras, run the following:

  port upgrade git-extras
```

###### Type(s)

- [x] bugfix
- [ ] enhancement
- [ ] security fix

###### Tested on

macOS 13.5.2 22G91 arm64
Xcode 14.3.1 14E300c

###### Verification
Have you

- [x] followed our [Commit Message Guidelines](https://trac.macports.org/wiki/CommitMessages)?
- [x] squashed and [minimized your commits](https://guide.macports.org/#project.github)?
- [x] checked that there aren't other open [pull requests](https://github.com/macports/macports-ports/pulls) for the same change?
- [x] checked your Portfile with `port lint --nitpick`?
- [x] tried a full install with `sudo port -vst install`?
- [x] tested basic functionality of all binary files?