openssl11: Update to 1.1.1w
This is the last release in the 1.1 series, so I'm adding a note to
highlight that this port must now be considered out of support.

See https://www.openssl.org/news/secadv/20230908.txt for the advisory.
The CVE does not affect macOS, but there are some other minor bug fixes
in the release.

CVE: CVE-2023-4807