qgis3{-ltr}: change linking from MacPorts' to the System's  libunwind
QGIS does not use libunwind, but due to it's dependence to PDAL (which does use libunwind)
and the following presence of libunwind in $prefix/lib, causes the MacPorts version to be picked
up by the linker instead of the system's version. A thrown exception is therefore wrongfully handled
by both versions, causing crash.

This is a workaround, awaiting separate prefix for the port libunwind.
See https://trac.macports.org/ticket/66250#comment:1

Closes https://trac.macports.org/ticket/68250