clang-14: Fix build under Xcode 15, re-enabling sanitizers on macOS 14
This also reverts the following as they relate to clang-14:

370d81aeeea1 llvm-14: Allow build on darwin23 but disable sanitizers
0c2af4734924 llvm-{10-14}: Restrict to darwin < 23

Note that the actual incompatibility was not strictly between clang-14
and Darwin 23 (macOS 14) as these previous commits suggest. The
incompatibility existed between clang-14 and the macOS 14 SDK, which is
part of Xcode 15. The incompatibility equally affected clang-14 on macOS
13 (Darwin 22) using Xcode 15.

This also makes the compiler-rt build for various runtimes targeting iOS
and iOS Simulator work correctly by removing the macOS SDK -syslibroot
when linking, allowing the platform-specific SDK provided by the Clang
build to be effective.

Fixes: https://trac.macports.org/ticket/68257