lilypond-devel: upgrade to 2.25.8
Note that '+docs' is currently broken.  However, it doesn't make sense to
fix this version because in the soon to be released version 2.25.9 the
ancient `texi2html` script (for which there never was a port in MacPorts)
will be no longer needed, which means that a thorough revision of this
Portfile is in order to build the complete documentation (and not just the
info files).